Python 3.4 compatibility
