+++ pymarsys/base_endpoint.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from abc import ABC, abstractmethod
 from .connections import BaseConnection
