+++ pymarsys/connections/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+import sys
+
+from .base import *
+from .exceptions import *
+from .sync import *
+
+if sys.version_info >= (3, 5):
+    from .async import *
