+++ pymarsys/connections/exceptions.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+__all__ = [
+    'ApiCallError',
+]
+
+
+class ApiCallError(Exception):
+    pass
