+++ pymarsys/contact_field.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from .base_endpoint import BaseEndpoint
