+++ tests/test_base_endpoint.py
@@ -1,9 +1,13 @@
+# -*- coding: utf-8 -*-
+
+from unittest import TestCase
+
 import pytest
 
 from pymarsys.base_endpoint import BaseEndpoint
 
 
-class TestBaseEndpoint:
+class TestBaseEndpoint(TestCase):
     def test_init(self):
         with pytest.raises(TypeError) as excinfo:
             BaseEndpoint()
