+++ tests/test_emarsys.py
@@ -1,5 +1,8 @@
-from pymarsys.contact import Contact
+# -*- coding: utf-8 -*-
+from unittest import TestCase
+
 from pymarsys.connections import SyncConnection
+from pymarsys.contact import Contact
 from pymarsys.emarsys import Emarsys
 
 EMARSYS_URI = 'https://api.emarsys.net/'
@@ -7,7 +10,7 @@ TEST_USERNAME = 'test_username'
 TEST_SECRET = 'test_secret'
 
 
-class TestEmarsys:
+class TestEmarsys(TestCase):
     def test_init_no_exception(self):
         connection = SyncConnection(TEST_USERNAME, TEST_SECRET)
         client = Emarsys(connection)
